+++ docs/_pages/releases.md
@@ -9,6 +9,7 @@ sidebar:
 ## Unreleased
 
 ### Fixes
+* Add missing reason to some equivalency assertions - [#405](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/405)
 * Add missing `StringSyntaxAttribute` to the because parameter of several assertions - [#404](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/404)
 * Fix assembly loading for NUnit4 and `LifeCycle.InstancePerTestCase` with Microsoft.Testing.Platform - [#362](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/362)
 * Fix configuring displayed length of string assertions at `AssertionScope` - [#393](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/393)